Generate table with coloured cells indicating expression values for selected genes

Generates table with coloured cells indicating expression values for selected genes.

exprTable(
  genes,
  keep_all = FALSE,
  data,
  cn_data = NULL,
  sv_data = NULL,
  cn_decrease = TRUE,
  targets,
  sampleName,
  int_cancer,
  ext_cancer,
  comp_cancer,
  add_cancer = NULL,
  genes_annot = NULL,
  oncokb_annot = NULL,
  cancer_genes = NULL,
  mut_annot = NULL,
  fusion_genes = NULL,
  ext_links = FALSE,
  type = "z",
  scaling = "gene-wise",
  caner_genes_annot.list
)

Arguments

genes

Selected genes.

keep_all

keep all rows

data

Input data.

cn_data

Copy number data.

sv_data

SV data.

cn_decrease

Order of the CN data.

targets

Target groups.

sampleName

Sample name.

int_cancer

Internal cancer group.

ext_cancer

External cancer group.

comp_cancer

Complete cancer group.

add_cancer

Additonal cancer type.

genes_annot

Geners annotation.

oncokb_annot

OncoKb annotation.

cancer_genes

Cancer genes.

mut_annot

Mutation annotation.

fusion_genes

Fusion genes.

ext_links

External links.

type

Type.

scaling

Scaling

caner_genes_annot.list

The caner_genes_annot.list from the report.

Value

Table with coloured cells indicating expression values for selected genes