Read WGS Data

Reads WGS data, including PCGR tiers.tsv, PURPLE cnv.gene.tsv, and Manta manta.tsv. If the file path has been specified in the RNAsum params and is valid, it is returned. As a fallback, if the umccrise directory param has been specified, then there is an attempt to detect the file pattern in there.

Usage

read_wgs_data(p)

Arguments

p

RNAsum params list.

Value

A list of the input sample data.

Examples

p <- list(
  umccrise = system.file("rawdata/test_data/umccrised/test_sample_WGS", package = "RNAsum"),
  pcgr_tiers_tsv = system.file(
    "rawdata/test_data/umccrised/test_sample_WGS/small_variants",
    "TEST-somatic.pcgr.snvs_indels.tiers.tsv",
    package = "RNAsum"
  ),
  manta_tsv = system.file(
    "rawdata/test_data/umccrised/test_sample_WGS/structural/TEST-prioritize-manta.tsv",
    package = "RNAsum"
  )
)
(res <- read_wgs_data(p))
#> $pcgr_tiers_tsv
#> # A tibble: 99 × 61
#>    CHROM POS       REF   ALT   GENOMIC_CHANGE    GENOME_VERSION VCF_SAMPLE_ID   
#>    <chr> <chr>     <chr> <chr> <chr>             <chr>          <chr>           
#>  1 3     179218303 G     A     3:g.179218303G>A  grch38         SBJ02999__PRJ22…
#>  2 17    7674894   G     A     17:g.7674894G>A   grch38         SBJ02999__PRJ22…
#>  3 13    110715614 C     G     13:g.110715614C>G grch38         SBJ02999__PRJ22…
#>  4 13    110719444 G     C     13:g.110719444G>C grch38         SBJ02999__PRJ22…
#>  5 8     20250093  C     T     8:g.20250093C>T   grch38         SBJ02999__PRJ22…
#>  6 8     13086457  C     T     8:g.13086457C>T   grch38         SBJ02999__PRJ22…
#>  7 10    31521795  G     C     10:g.31521795G>C  grch38         SBJ02999__PRJ22…
#>  8 9     132927250 G     A     9:g.132927250G>A  grch38         SBJ02999__PRJ22…
#>  9 3     47121452  G     A     3:g.47121452G>A   grch38         SBJ02999__PRJ22…
#> 10 6     117365075 G     C     6:g.117365075G>C  grch38         SBJ02999__PRJ22…
#> # ℹ 89 more rows
#> # ℹ 54 more variables: VARIANT_CLASS <chr>, SYMBOL <chr>, GENE_NAME <chr>,
#> #   CCDS <chr>, CANONICAL <chr>, ENTREZ_ID <chr>, UNIPROT_ID <chr>,
#> #   ENSEMBL_TRANSCRIPT_ID <chr>, ENSEMBL_GENE_ID <chr>, REFSEQ_MRNA <chr>,
#> #   ONCOSCORE <chr>, ONCOGENE <chr>, TUMOR_SUPPRESSOR <chr>,
#> #   ONCOGENE_EVIDENCE <chr>, TUMOR_SUPPRESSOR_EVIDENCE <chr>,
#> #   DISGENET_CUI <chr>, DISGENET_TERMS <chr>, CONSEQUENCE <chr>, …
#> 
#> $purple_gene_tsv
#> # A tibble: 25,417 × 20
#>    chromosome  start    end gene        minCopyNumber maxCopyNumber unused
#>    <chr>       <int>  <int> <chr>               <dbl>         <dbl> <chr> 
#>  1 chr1        11869  14409 DDX11L1              1.98          1.98 0     
#>  2 chr1        14404  29570 WASH7P               1.98          1.98 0     
#>  3 chr1        17369  17436 MIR6859-1            1.98          1.98 0     
#>  4 chr1        29554  31097 MIR1302-2HG          1.98          1.98 0     
#>  5 chr1        30366  30503 MIR1302-2            1.98          1.98 0     
#>  6 chr1        34554  36081 FAM138A              1.98          1.98 0     
#>  7 chr1        69091  70008 OR4F5                1.98          1.98 0     
#>  8 chr1       185217 195411 FO538757.1           1.98          1.98 0     
#>  9 chr1       187891 187958 MIR6859-2            1.98          1.98 0     
#> 10 chr1       450740 451678 OR4F29               1.98          1.98 0     
#> # ℹ 25,407 more rows
#> # ℹ 13 more variables: somaticRegions <dbl>, germlineHomDeletionRegions <dbl>,
#> #   germlineHetToHomDeletionRegions <dbl>, transcriptId <chr>,
#> #   transcriptVersion <chr>, chromosomeBand <chr>, minRegions <dbl>,
#> #   minRegionStart <int>, minRegionEnd <int>, minRegionStartSupport <chr>,
#> #   minRegionEndSupport <chr>, minRegionMethod <chr>,
#> #   minMinorAlleleCopyNumber <dbl>
#> 
#> $manta_tsv
#> $manta_tsv$unmelted
#> # A tibble: 99 × 19
#>    vcfnum  nann TierTop Start         End    Type  BND_ID BND_mate SR_alt PR_alt
#>    <chr>  <dbl> <chr>   <chr>         <chr>  <chr> <chr>  <chr>     <int>  <int>
#>  1 68         1 4       1:107,176,815 orpha… BND   1      B             7      9
#>  2 69         1 4       1:107,177,428 orpha… BND   2      A            23     26
#>  3 76         1 4       1:114,885,352 orpha… BND   3      A            27     35
#>  4 01         1 4       1:2,652,000   1:NA   PURP… NA     NA           NA     NA
#>  5 02         1 4       1:3,614,001   1:NA   PURP… NA     NA           NA     NA
#>  6 03         1 4       1:6,593,001   1:NA   PURP… NA     NA           NA     NA
#>  7 04         1 4       1:6,908,000   1:NA   PURP… NA     NA           NA     NA
#>  8 05         1 4       1:7,714,001   1:NA   PURP… NA     NA           NA     NA
#>  9 06         1 4       1:10,615,000  1:NA   PURP… NA     NA           NA     NA
#> 10 07         1 4       1:12,123,001  1:NA   PURP… NA     NA           NA     NA
#> # ℹ 89 more rows
#> # ℹ 9 more variables: SR_PR_sum <int>, SR_PR_ref <chr>, Ploidy <dbl>,
#> #   AF_PURPLE <chr>, AF_BPI <chr>, CNC <chr>, CN <chr>, SScore <int>,
#> #   annotation <chr>
#> 
#> $manta_tsv$melted
#> # A tibble: 186 × 28
#>    vcfnum  nann TierTop Start         End    Type  BND_ID BND_mate SR_alt PR_alt
#>    <chr>  <dbl> <chr>   <chr>         <chr>  <chr> <chr>  <chr>     <int>  <int>
#>  1 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  2 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  3 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  4 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  5 62         5 3       1:97,924,845  1:97,… DUP   NA     NA           21     15
#>  6 13         4 3       1:20,551,467  1:20,… DEL   NA     NA           11     21
#>  7 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  8 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#>  9 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#> 10 82        61 3       1:150,364,461 1:150… DUP   NA     NA           18     28
#> # ℹ 176 more rows
#> # ℹ 18 more variables: SR_PR_sum <int>, SR_PR_ref <chr>, Ploidy <dbl>,
#> #   AF_PURPLE <chr>, AF_BPI <chr>, CNC <chr>, CN <chr>, SScore <int>,
#> #   Event <chr>, Effect <chr>, Genes <chr>, Transcript <chr>, Detail <chr>,
#> #   Tier <chr>, ntrx <dbl>, ngen <dbl>, neff <dbl>, `Tier (Top)` <glue>
#> 
#> $manta_tsv$tsv_descr
#> # A tibble: 23 × 3
#>    Column             Description                                          Type 
#>    <chr>              <chr>                                                <chr>
#>  1 caller             Manta SV caller                                      c    
#>  2 sample             Tumor sample name                                    c    
#>  3 chrom              CHROM column in VCF                                  c    
#>  4 start              POS column in VCF                                    i    
#>  5 end                INFO/END: End position of the variant described in … i    
#>  6 svtype             INFO/SVTYPE: Type of structural variant              c    
#>  7 split_read_support FORMAT/SR of tumor sample: Split reads for the ref … c    
#>  8 paired_support_PE  FORMAT/PE of tumor sample: ??                        c    
#>  9 paired_support_PR  FORMAT/PR of tumor sample: Spanning paired-read sup… c    
#> 10 AF_BPI             INFO/BPI_AF: AF at each breakpoint (so AF_BPI1,AF_B… c    
#> # ℹ 13 more rows
#> 
#> $manta_tsv$col_descr
#> # A tibble: 30 × 2
#>    Column     Description                                                       
#>    <chr>      <chr>                                                             
#>  1 nrow       "Row number that connects variants between tables in same tab set…
#>  2 vcfnum     "Original event row number that connects variants to events."     
#>  3 TierTop    "Top priority of the event (from simple_sv_annotation: 1 highest,…
#>  4 Tier       "Priority of the specific event (from simple_sv_annotation: 1 hig…
#>  5 Tier (Top) "Priority of the specific event (Top tier of all event's annotati…
#>  6 Chr        "Chromosome."                                                     
#>  7 Start      "Start position as inferred by BPI. For PURPLE-inferred SVs we us…
#>  8 End        "End position. For BNDs = Chr:Start of the mate.Values are inferr…
#>  9 ID         "ID of BND from Manta. For PURPLE-inferred SVs this is PURPLE."   
#> 10 MATEID     "ID of BND mate from Manta."                                      
#> # ℹ 20 more rows
#> 
#>